requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.9
django-adminactions==2.1.0
django-admin-list-filter-dropdown==1.0.3
django-autocomplete-light==3.9.7
django-bootstrap4==21.1
django-clearcache==1.2.1
django-crispy-forms==1.11.2
django-extensions==3.1.5
django-filter==21.1
django-mathfilters==1.0.0
django-partial-date==1.3.1
django-tables2==2.4.1
dj-database-url==2.0.0
entrypoints==0.3
google-cloud-translate
gunicorn==20.1.0
html5lib==1.1
httpx
idna==3.3
importlib-resources==5.4.0
ipykernel==6.23.1
ipython<8
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
joblib==1.2.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.4.4
jupyter-console==6.4.0
jupyter-core==5.3.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jupytext==1.13.7
kiwisolver==1.3.2
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.8
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.4
nltk==3.8.1
notebook==6.5.4
numpy==1.24.3
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pbr==5.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pyzmq==24
qtconsole==5.2.0
QtPy==1.11.2
regex==2021.11.10
requests==2.26.0
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
spotipy==2.23.0
sqlparse==0.4.4
terminado==0.12.1
testpath==0.5.0
testresources==2.0.1
textblob==0.17.1
tornado==6.3.2
tqdm==4.62.3
traitlets==5.6.0
Unidecode==1.3.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
